requirements.txt
Django>=2.2,<3.0
psycopg2>=2.7,<3.0

django-cms>=3.7,<3.8

django-filer>=1.6,<1.7
djangocms-text-ckeditor>=3.8,<3.9
djangocms-link>=2.5,<2.6
djangocms-file>=2.3,<2.4
djangocms-picture>=2.3,<2.4
djangocms-video>=2.2,<2.3
djangocms_googlemap>=1.3,<1.4
djangocms_snippet>=2.2,<2.3
djangocms_style>=2.2,<2.3